requirements.txt
boto==2.29.1
python-gflags==2.0
requests==2.3.0
